requirements.txt
python-telegram-bot[ext]
google-generativeai
python-dotenv
